修改一些bug；球员详情的赛季比赛记录；
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@645277 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/nba_player_season_game_entity.dart
@@ -29,7 +29,7 @@ class NbaPlayerSeasonGameEntity {
 	late double teamId;
 	late double threePm;
 	late double to;
-	late double awayTeamId;
+	late int awayTeamId;
 
 	NbaPlayerSeasonGameEntity();
